test refactor/breakdown
